backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
eth-account>=0.10.0
web3>=6.0.0
pydantic>=2.0.0
requests>=2.28.0
